PP-4205 Extract logic to decide if charge is cancellable out of return_controller
- this logic moved because originally we wanted to re-use this code,
  however we changed our implementation for direct redirect to not do
  any cancellation
- we are keeping this change because we think it is an improvement
  regardless

with @DanailMinchev
